requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-sqlalchemy==3.0.5
werkzeug==2.3.7
pypdf2==3.0.1
langchain==0.0.267
requests==2.31.0
python-dotenv==1.0.0